Hopefully working example
Added the build.sbt with the tut dependency, the plugins file with tut
added, and an example src and target file. README is pointed at the
generated file, which is ok. I don't like the idea of checking in
target/ files, so that's a problem that would have to be solved. I
experimented some with site as well, but couldn't get it to generate
anything. I'll hammer more at it some other time.